requirements.txt
# Core runtime
onnxruntime>=1.17.0          # CPU inference (auto-detects GPU via onnxruntime-gpu)
numpy>=1.24.0
opencv-python-headless>=4.8  # headless = no GUI deps on servers
Pillow>=10.0

# Similarity search
faiss-cpu>=1.7.4             # swap to faiss-gpu if CUDA available

# Clustering
hdbscan>=0.8.33              # better than DBSCAN for unknown cluster count
scikit-learn>=1.3.0          # fallback + preprocessing

# Utilities
tqdm>=4.66                   # progress bars for batch processing
scipy>=1.11                  # similarity transform math